+++ app/javascript/packs/survey.js
@@ -1,8 +1,8 @@
 import Vue from 'vue';
 import Vuelidate from 'vuelidate';
 import VueI18n from 'vue-i18n';
-import App from '../service/App.vue';
-import i18n from '../service/i18n';
+import App from '../survey/App.vue';
+import i18n from '../survey/i18n';
 
 Vue.use(VueI18n);
 Vue.use(Vuelidate);
@@ -18,7 +18,7 @@ window.bus = new Vue();
 Vue.config.productionTip = false;
 
 window.onload = () => {
-  window.WOOT_SERVICE = new Vue({
+  window.WOOT_SURVEY = new Vue({
     i18n: i18nConfig,
     render: h => h(App),
   }).$mount('#app');
