+++ app/javascript/survey/assets/scss/woot.scss
@@ -16,6 +16,6 @@ body {
   height: 100%;
 }
 
-.woot-service-wrap {
+.woot-survey-wrap {
   height: 100%;
 }
