+++ app/javascript/survey/components/Rating.vue
