+++ app/javascript/survey/i18n/index.js
