+++ app/views/survey/responses/show.html.erb
@@ -7,8 +7,8 @@
     <script>
       window.globalConfig = <%= raw @global_config.to_json %>
     </script>
-    <%= javascript_pack_tag 'service' %>
-    <%= stylesheet_pack_tag 'service' %>
+    <%= javascript_pack_tag 'survey' %>
+    <%= stylesheet_pack_tag 'survey' %>
   </head>
   <body>
     <div id="app"></div>
