+++ spec/controllers/service/responses_controller_spec.rb
@@ -1,15 +1,15 @@
 require 'rails_helper'
 
-describe '/service/response', type: :request do
-  describe 'GET service/responses/{uuid}' do
+describe '/survey/response', type: :request do
+  describe 'GET survey/responses/{uuid}' do
     it 'renders the page correctly when called' do
       conversation = create(:conversation)
-      get service_response_url(id: conversation.uuid)
+      get survey_response_url(id: conversation.uuid)
       expect(response).to be_successful
     end
 
     it 'returns 404 when called with invalid conversation uuid' do
-      get service_response_url(id: '')
+      get survey_response_url(id: '')
       expect(response.status).to eq(404)
     end
   end
