+++ tailwind.config.js
@@ -4,7 +4,7 @@ module.exports = {
   purge: [
     './app/javascript/widget/**/*.vue',
     './app/javascript/shared/**/*.vue',
-    './app/javascript/service/**/*.vue',
+    './app/javascript/survey/**/*.vue',
   ],
   future: {
     removeDeprecatedGapUtilities: true,
